reorganize client folders, pull components out of app.js
