+++ client/src/pages/login.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { compose } from 'react-apollo';
-import LoginForm from './LoginForm';
+import LoginForm from '../components/LoginForm';
 
 class Login extends Component {
   render() {
